requirements.txt
openai-whisper
tqdm 